Initial design for support of HMAC precomputed keys
This commit implements an optimization of HMAC described in
NIST-FIPS-198-1 (Section 6) and in RFC2104 (Section 4).

Instead of using the original key, it is possible to use a precomputed
key consisting of two intermediate hashing results. Using this
precomputed key removes at least two calls to the hash compression
function when computing a HMAC value.

As an initial design, only SHA-256 is supported and all other hashes are
disabled from HMAC. This makes some unit tests fail. Once this design is
validated, we will add back the other hash functions.

To allow partial testing, we updated hmac_tests.txt to only contain the
HMAC-SHA256 tests. The original file is hmac_tests_full.txt

Testing can be done in the CMake build folder using:

* `crypto/crypto_test --gtest_filter='*InitAndGetStateSHA256*:HMACTest.TestVectors'`
  for the new/changed unit tests
* `ninja acvp_tests` for the ACVP tests (failure is expected for
  anything using HMAC-xx where xx is not SHA-256)